Update patch set 1
Patch Set 1: Run-TryBot+1

To be honest, I'm not sure why dtoi supports negative numbers in the first place. After this CL, the only usages of it are in dnsconfig_unix.go where its not obvious that signed integers need to be supported.

Is anyone opposed to just removing the logic to handle negative numbers?

Patch-set: 1
Label: Run-TryBot=+1