Update patch set 2
Patch Set 2:

(2 comments)

LGTM. I'll +2 after the rename of bitMarked.

> This CL does mean that we needlessly load hbits in some cases (otherwise, indeed, the compiler would have done this for us), but that's (apparently) an acceptable trade-off.

Is that really true? If i == sys.PtrSize, we have to load it for isPointer. If i != sys.PtrSize, we have to load it for morePointers (and may also have to load it for isPointer). So we have to load it at least once in the current code, and we load it exactly once in the new code.

Patch-set: 2