cmd/compile: add printfix format checker tool
The printfix command is a utility to enumerate, verify,
and automatically rewrite format strings for the package
in this directory.

printfix finds all potential (Printf, etc.) format strings.
If they are used in a call, the format verbs are verified
based on the matching argument type against a precomputed
table of valid formats (formatMapping, below). The table
can be used to automatically rewrite format strings with
the -u flag.

Run as: go run printfix.go {flags}

Flags:
	-t  test that format specifiers are valid
	-u  update format specifiers per formatMapping
	-v  verbose mode

A new formatMapping can be generated by running in -v mode.

This tool rather specialized at the moment but will help
ensure that we use the correct format strings throughout
the compiler, especially once we switch to using fmt.Formatter.

Change-Id: I1ca5bb6421d57ac78a00f1a80b9547a72837adc9