runtime: remove maxstring
The runtime prevents printing of overlong strings with the print function when
the length of the string is determined to be corrupted. Corruption is checked by
comparing the string size against the limit that is stored in maxstring.

However maxstring does not get updated everywhere were go strings are created
e.g. for string constants during compile time. Thereby the check for maximum
string length prevents the printing of some valid strings.

The protection maxstring provides does not warrant the bookkeeping
and global synchronization needed to keep maxstring updated to the
correct limit everywhere.

Fixes #16999

Change-Id: I62cc2f4362f333f75b77f199ce1a71aac0ff7aeb