cmd/compile: prototype/experiment for register args and results
DO NOT SUBMIT

This is work in progress for passing arguments (parameters?
what is the proper word?) and results in registers.

Moving parts planned:
A. a pragma for indicating register-ness of a function
   for now this is for experiments, in the future it will
   have the opposite sense for legacy assembly language
B. generic Load/Store instructions signaling use of registers
C. architecture-specific "fake" load/store instructions to
   allow plive unmodified (these will be stripped out after
   plive) and to work smoothly with regalloc.
D. logic to trigger use of appropriate
   instructions for registers.  This falls into 4 parts:
   a. pass in registers
   b. receive in registers
   c. return out registers
   d. receive out registers
E. annotating dwarf appropriately.
F. automatically wrapping legacy assembly language functions
   when they are passed as function pointers or used as
   interface methods.

This CL includes prototype versions of A-C, and D.a.
Looking at D.b and the parameter return handshake from
*state.call in gc/ssa.go makes me think that I should create
a register-or-memory type for decladdrs and the result from
*state.call, which is where I'm at now.  One difficulty is
that multiword values (slices, strings, interfaces) need to
mention multiple registers, so it's not just "a register".

Concerns:
- I don't like that this essentially depends on the kindness
  of schedulers; there's nothing else tying the calls to the
  "fake" instructions that define inputs and receive their
  results.  I don't see any promises in the design of the
  register allocator that it wouldn't arbitrarily stomp on a
  parameter/result register while shuffling things around.
- Prototype uses 3 registers, I think we're aiming for 6 or
  more.  This might get tedious and verbose, there might be
  scaling problems in the code (e.g., where before we had
  OpStore, now we have a dozen store opcodes).
- Haven't done the work to consider callee-saved registers
  (yet).  That might have some common solution with the first
  concern.

Change-Id: Ic4598c3fbb66a3bf8805fe27ed023f34e91e80dc