cmd/compile: prototype/experiment for register args and results
DO NOT SUBMIT

(rebased)

Experimentally added PPC64, works (or not) same as AMD64.
Experimentally added ARM64, works.
Sync'd w/ new SSA assignment, fixed bug w/ addressed params.
Sync'd w/ clean up in ssa.go
1% reduction in size for runtime, code runs generally faster
but sometimes slower, so widening coverage of opt-in for
benchmarking to see if it yields expected improvements.

This is work in progress for passing arguments (parameters?
what is the proper word?) and results in registers.

Moving parts planned:
A. a pragma for indicating register-ness of a function
   for now this is for experiments, in the future it will
   have the opposite sense for legacy assembly language
B. generic Load/Store instructions signaling use of registers
   (Store instructions currently used on caller side, callee
    side uses specialized "Arg" ipcodes instead.)
C. architecture-specific "fake" load/store instructions to
   allow plive unmodified (these will be stripped out after
   plive) and to work smoothly with regalloc.
   (These appear to not be necessary for passing parameters;
   instead, the Arg stores and Arg ops code-generate to
   nothing; existing conventions for liveness appear to work,
   certainly on the caller side)
D. logic to trigger use of appropriate
   instructions for registers.  This falls into 4 parts:
   a. pass in registers
   b. receive in registers
   c. return out registers
   d. receive out registers
E. annotating dwarf appropriately.
F. automatically wrapping legacy assembly language functions
   when they are passed as function pointers or used as
   interface methods.

This CL includes prototype versions of A-C, and D.a-b.
The caller-side passes in registers according to rules,
the callee-side receives in registers according to rules,
but it is not clear that the liveness information is right,
and the ASM-generated stack-growth code needs to spill and
reload.

Works, with limits:
Exported functions work.
Will FAIL for function pointers.
Will FAIL for methods that are interface methods.
Will FAIL for reflective calls.

For strings, slices, and interfaces, it appears that doing
part of the user-level decompose early would help; that is,
the pieces pass in args, fed to a Make String/Slice/Interface.

For the simplest and earliest experiment, the callee might
aggressively-always store parameters into spill locations,
which is correct against stack growth.  Notably, this
achieves the calling convention.

Concerns:
- I don't like that this essentially depends on the kindness
  of schedulers; there's nothing else tying the calls to the
  "fake" instructions that define inputs and receive their
  results.  I don't see any promises in the design of the
  register allocator that it wouldn't arbitrarily stomp on a
  parameter/result register while shuffling things around.

  Update: on the caller side, these are "stores", with a memory
  operand to help force an order, and the "stores" are also
  performed in registers-last order.  Modifying the register
  allocator to mark the parameter registers nospill between
  the "store" operations and the next call also guarantee
  the desired behavior (but this has not yet been done).

- Prototype uses 3 registers, I think we're aiming for 6 or
  more.  This might get tedious and verbose, there might be
  scaling problems in the code (e.g., where before we had
  OpStore, now we have a dozen store opcodes).

- Haven't done the work to consider callee-saved registers
  (yet).  That might have some common solution with the first
  concern.

Change-Id: Ic4598c3fbb66a3bf8805fe27ed023f34e91e80dc