Update patch set 8
Patch Set 8:

With regards to the comments on patch set 8. I think it's helpful to step back and have a discussion as to why this feature is requested. The use case I have in mind (I think this is the common use case) is that json input is being received from an unreliable source. I would like to know when there are json keys that are not being populated into struct fields and which keys were not used.

decode.go:121. This could be changed to one bad field. I believe it's a more helpful message to report all the fields that are in error.

decode.go:694. These are not duplicate fields. These are json keys that have collided according to the specification of the json unmarshaller. See for example https://github.com/golang/go/blob/master/src/encoding/json/decode_test.go#L537. The 'X' and the 'x' keys have collided and therefore neither of the values are populated into the struct.

decode.go:704. In my opinion reporting the context of the nested object is out of scope. It is usually sufficient to report the names of the keys.

decode_test.go:382. This is not overkill. For the large majority of the test cases in the unit tests, the unmarshal output is idempotent to the DisallowUnknownFields() flag. However in a few cases the output is different depending on the flag. In those cases I have added tests that check the expected output on the two cases.

encode.go:1256. Yes and no for this change is unrelated. I must ensure that the order of the output array in cachedTypeFields() is preserved.  This can be done by adding the 2nd read of the value while holding the lock. Or I can change decode.go:669 and move the call to cachedTypeFields() outside of the loop. IMO the change I have made is simpler but I can do the 2nd approach.

Patch-set: 8
Label: Code-Review=0