Update patch set 1
Patch Set 1:

An additional concern with the memory-or-register abstraction is how we deal with mixed-size data types -- pointers and integer on amd64p32, and single-vs-double precision registers on some architectures with overlapping register sets (e.g., Sparc V8, where F0+F1 are aliased to D0, etc.).  One problem example for 64p32 is a function returning an int and and error interface -- so one 64 bit register containing 64 bits of int, one 64-bit register containing 32 bits of pointer for the interface, and another 64-bit register containing 32 bits of type information for the interface.

Patch-set: 1