Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > Is 4% really worth it for the added complexity? (It's not much
 > complexity, but ... 4%)
 > 
 > Does this speed up anything that people care about? Do people care
 > about ScanInts?

It speeds up many fmt scans that use bytes.buffer a little since UnreadRune gets a lot faster. But UnreadRune is only a small part of fmt scan.

I think there is also an issue with the old code (which i just realized) that when reading a rune error in an invalid utf8 sequence it returns size 1 and advances 1 but when we unread we actually unread a fullrune thereby we could end up in a position before readrune. So Unreadrune will not necessary unread the last read byte sequence/rune. But that might actually be intended.
https://play.golang.org/p/nmTjktpSdr
is this an issue in need of a fix?

I am fine with abandoning if a small percent increase for e.g. fmt scan does not seem worth the added complexity and this CL does not fix any error.

Patch-set: 2
Label: Code-Review=0