Update patch set 4
Patch Set 4:

what happens when the url struct contains "mongodb://host1,host2:27017,host3:27018/" ?

as mentioned in #12023, the package doesn't comply with rfc 3986, and has no concrete definition of the authority component syntax. perhaps, it might be better to leave this package as it is, instead have a new url package that conforms to rfc 3986 and copy over the vendor folder.

Patch-set: 4