Update patch set 6
Patch Set 6:

All done.

Unfortunately, in the interim, there are new errors that will take some work to fix, since the right fix is to vet--see issue 17057.

Since this is not yet wired up to all.bash or the builders, however, I'm going to submit it. Any incremental changes will be easier to review in a new CL.

Patch-set: 6
Label: Code-Review=0