Update patch set 8
Patch Set 8: Code-Review+1

(2 comments)

I think the operand ordering is okay (where I looked, it appeared okay, need to check it for vector compares but I assume the author already knows the answer to that).  The translation order needs to be documented so that whoever goes to use this does not either determine it by experiment (done that once or twice) or by laboriously tracing the operand flow through the assembler source while cross referencing to the bits in the ASIrewoP (done that three times, once correctly).

Patch-set: 8
Label: Code-Review=+1