Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> From the ISA:
> isel rt,ra,rb,bc // bc is a constant
> rt@6 ra@11 rb@16 bc@21
> 
> fsel frt,fra,frc,frb // all fregs; same as fmadd and others
> frt@6 fra@11 frb@16 frc@21
> 
> It has been my assumption that the go asm operand order is the opposite of the power asm operand order.  I'm assuming too that the fsel is correct since it is the same as the fmadd and other 4 operand FP instructions and those probably work.

I still need to review my reply above with the bit order done correctly, but the Go order is not necessarily just the reverse.  Often the input operands remain in the same order, but the target is moved from left to right.  That seems to be the case for the vector ops CL I recently left a comment on (and that is good).  Otherwise, non-commutative operations end up backwards and people (and compiler writers) get confused.

Patch-set: 1