Update patch set 3
Patch Set 3:

> Just need a comment on runtime/string.go:slicebytetostringtmp
 > saying it only gets used when instrumenting.

That would not be correct since only OARRAYBYTESTRTMP generated by the frontend are intrinsified. I addmit that in that regard the CL name is wrong and should be changed. 

However with the old backend being deleted in mind i can change the CL to:
use generate slicebytetostringtmp calls for OARRAYBYTESTRTMP in frontend unconditionally (like before). In ssa use the intrinsic call infrastructure to replace the call when not instrumenting by optimized opcodes and otherwise leave the call as is. This way there is no OARRAYBYTESTRTMP after the frontend and runtime calls get intrinsified too. 

The comment and CL name would then be accurate for the ssa backend but not for the old backend.

Patch-set: 3