Update patch set 3
Patch Set 3:

> i just tried to teach the ssa backend to intrinsify runtime
 > slicebytetostringtmp by intrinsicCall. However it seems never
 > triggered for runtime since i added intrinsicKey{"runtime",
 > "slicebytetostringtmp"} but when slicebytetostringtmp is found
 > sym.Pkg.Path is empty. Does runtime always have and is the only one
 > to have "" as Pkg.Path?

Something more generic in runtime/string.go then.  I just want to reader to realize that this may not always be called even though it appears to be called in the source.

pkg.Path=="" means the package currently being compiled.  The other intrinsifications are all cross-package.  This one is intra-package,so it is a bit different.

Patch-set: 3