Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> Regarding documenting the translation order, would comments in the code suffice?
> 
> I'll work on the other changes you suggested.
> 
> Thanks

Comments in the code is about all we have for anything else, so that's fine.
I may complain about your choices, but at least we'll agree on what they are.

I think it's good to have a little redundancy, maybe a quick description of the semantics plus a mapping to the PPC instruction and register names (e.g., FRA, FRB, VRA, VRB, etc), because there's so much potential for confusion here (e.g., assembly language can be confusing if the PPC asm operand order is different from the instruction order, as is the case with fmadd).  I'd like to include the instruction name because of the slim but not-zero possibility that in the future we might try to babelize the common parts of all these architectures, and then it wouldn't be a simple up/down case mapping.

So e.g., 
// VCMPGTSB x,y,d == for byte vectors x,y,d, 0<=i<8, d[i]=x[i]>y[i]?1:0 == vcmpgtsb VRA=x, VRB=y, VRT=d

Patch-set: 8