Update patch set 2
Patch Set 2:

Perhaps more helpful is a verbose mode that gives
the exact reason why a source file (*.go, *.c, *.s,
etc.) is ignored.

Even though the "no buildable Go files" error might
be confusing at first, the user should be quick to
learn that it's because files are ignored. The real
tricky thing is to figure out why a certain file is
ignored, IMHO.

Patch-set: 2