Update patch set 2
Patch Set 2:

> Perhaps more helpful is a verbose mode that gives
 > the exact reason why a source file (*.go, *.c, *.s,
 > etc.) is ignored.

We can do that but I fear that the output will be very verbose and sometimes hard to read.

I wonder if it makes sense to report the ignored files and why if -v is set.
 
 > Even though the "no buildable Go files" error might
 > be confusing at first, the user should be quick to
 > learn that it's because files are ignored. The real
 > tricky thing is to figure out why a certain file is
 > ignored, IMHO.

Actually, we are doing nothing but to make the error message a little bit more understandable without having to stackoverflow the message to see what is means. Current messaging is learnable but adding a little bit more info makes it more self explanatory.

Patch-set: 2
Label: Code-Review=0