syscall: document Exec first parameter
When you run os.Exec("ls", "-al"), the os package will find the first ls binary
in the PATH, and run that. However, syscall.Exec("ls", "-al") will report "no
such file or directory" when run. I didn't expect this and the error didn't
make it clear how to fix it; my first steps involved modifying the PATH of the
containing program, and setting PATH in the third argument to syscall.Exec.

Document that syscall.Exec expects a full path to the executable
(Exec("/bin/ls", "-al"), and that it won't use any set PATH variable.

Change-Id: I611511434f37c75f77c22f61f469108243bc0101