Update patch set 3
Patch Set 3: Code-Review-1

This is all part of man 3 exec on Unix. I do not believe we need to get in the habit of documenting the workings of every system call like this.

Package syscall is an interface to the system calls and is very system-specific. That is its nature. We should let the system documentation stand; doing it ourselves is infeasible.

The only part that's documented now is the Go-specific pieces such as StringByteSlice. That's as it should be.

Patch-set: 3
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1