Update patch set 8
Patch Set 8:

(2 comments)

Thanks for the thorough explanation. It's not clear to me that we really want this many examples here. I'll need a second opinion.

But until we decide would you mind splitting it into two changes, so that we can get the obviously useful first two examples in?

Patch-set: 8
Label: Code-Review=0