Update patch set 1
Patch Set 1:

> Sorry, I didn't see this before I wrote CL 29248. Happy to abandon
 > that one in favor of this.
 > 
 > One question: why move RelocVariant into the obj package? It's only
 > used in cmd/link, so we can keep it there.

No worries David. Giving relocation variants a type is nice regardless of whether this proposal gets accepted or not.

The idea behind this CL is to allow the obj package to write relocation variants to object files, hopefully reducing the amount of architecture-specific code in the linker. The RelocVariant type therefore has to live in the obj package. Although I think there was talk of moving these types/constants to a third package so that ld didn't have to import obj.

My focus has been on SSA recently but I hope to revisit this soon.

Patch-set: 1