Update patch set 1
Patch Set 1:

I haven't seen any examples on custom marshalJSON and UnmarshalJSON, not in the documentation nor in the blog post https://blog.golang.org/json-and-go.

If this example doesn't fit here, please let me know where I can make it to.

I noticed that some folk struggle with custom JSON marshalling especially with the part where the values being passed in are quoted, and when marshaling the out value to ensure it is quoted. This deduction is from interacting with one or two others and having to explain the custom methods.

Patch-set: 1
Label: Code-Review=0