Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> What will happen to argument value dumps in
> stack traces if we pass some of them in
> registers?
> 
> What's the performance benefit of passing
> argument with registers vs. better inlining
> and function (partial-)specialization?

Claim, from two stack trace suppliers I talked to (Delve and Backtrace.io) is that they use Dwarf, so we just say what we're doing in Dwarf and life is good.  In practice I am less sure how that will work with SSA in the picture, unless we are careful to do tracking at a fine grain (and what if they are dead?).  But we also have that problem slightly now; if we modify the parameter in a register and don't need that register before the value goes dead, it won't store back, and the "wrong" (stale) value will appear.

Current plan also has exactly the same stack layout preserved for parameters, to be used by callee as their spill area, so one experiment will be performance with/without those spills always enabled in the SSA-generated prologue.  Those spills will also appear in the ASM-generated if-stack-overflow-or-safepoint code, in the event either occurs.

I expect we'll get a few percent, and why "vs"?  We can do both, compilers for other languages certainly do.  From a very contrived example:
- artificial/best-case 28% reduction in time (i.e., actual results will be worse)
- less-artificial 20% reduction in caller code size (i.e., will sometimes see this for real)

Patch-set: 4