Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> This change is a little sad. I'd hope that 10 small allocations were as fast as 1 10x-sized allocation. Then all 10 could have their own separate lifetimes, without pinning the memory of the other fields if only 1/10th of the memory is retained by the caller for a long period of time.
> 
> I realize there are 10x as many objects to count and track in that case, though.
> 
> Maybe this change is fine, since most CSV record lines will be small-ish.
> 
> Austin, thoughts?

It would be nice if 10 small allocations were the same as 1 10x-sized allocation, but that's not realistic. There is some small constant overhead to performing an allocation. These benchmarks, in particular, represent the worst possible case for the allocator and the current code, since nearly all of the allocations will be a single byte, or at most a few words. They also represent basically the best cast for the new code.

Justin, do you have performance numbers for more realistic CSV data that has longer field values?

Patch-set: 6