Update patch set 22
Patch Set 22:

Tried to use this today, and was reading the patches for the first time too. It's a lot of code to support only the client side of it. I noticed it implements Fast Open for Linux and Plan9 but not Darwin.

For the server side, I could not find a way to call setsockopt on net.TCPListener, so ended up using this reflection trick from https://github.com/aristanetworks/goarista/blob/master/dscp/listen_unix.go to get my tests going. Confirmed that it works with curl 0.49 and --tcp-fastopen. I had to define TCP_FASTOPEN=23 for Linux and TCP_FASTOPEN=0x105 for Darwin.

Patch-set: 22