runtime: preempt idle mark workers on any other work
DO NOT REVIEW. Experiment for #16528.

Currently idle mark workers only start if a P is idle, but they run
for a full scheduling quantum even if other work comes in before then.
This is particularly problematic in a network server, which tends to
have a lot of very short idle periods between network requests coming
in.

This is an experiment in making the idle mark worker constantly poll
for either scheduler or network work and immediately return if there
is other work to do.

This does in fact break up the idle worker quite well in the
go-gcbench/rpc benchmark; however, tail latency is still high. For
BenchmarkRPC/reqs-per-sec:10000/ballast:64MB/gomaxprocs:2, P99 latency
is 38ms and P99.9 is 57ms. The execution trace suggests that the
server falls behind roughly between the beginning of each GC cycle and
when background credit becomes consistently positive and assists stop
kicking in (this period is roughly 70ms).

There are some horrible hacks in here. I added fake sweep and mark
termination trace events around idle workers and assists,
respectively, just so I could identify them in the execution trace.

Change-Id: I30dbbbe0aa7e8e4312f47cd412b76ea218657833